Fixes for problems found when reviewing rules.md
This fixes the problems that I found while reviewing the documentation
page that is aiming at giving an introduction to the rule engine. Most
of the fixes are to improve the language and fix some small
inaccuracies. I also rewrote some of the examples as I found the
original ones difficult to understand.

Fixes:
https://emqx.atlassian.net/browse/EMQX-9186