Improvements to the description of the rule-engine SQL-like language
This commit contains the improvements I made while reviewing the
description of the SQL language. It is still possible to improve the
document; for example, one could describe the syntax of expressions more
precisely instead of just describing it implicitly by giving examples,
as we are currently doing. However, after the changes in this commit, I
think it will be much easier for the reader to understand the language
and what can be done with it, so I think it is a good start.