docs: fixes in rule-sql-builtin-functions.md
This commit include fixes that I made while reviewing the document. I
still think there are many improvements that could be made. For example,
it would be useful with a short introduction before every group of
functions.

The structure of the time and date functions could be improved. The
functions that takes an unit or format string as parameter should have
references to descriptions and tables with more detailed descriptions of
the available options.